{{cookiecutter.project_name}}/requirements.txt
raven>=3
asn1crypto==0.24.0
aspy.yaml==1.0.0
cached-property==1.4.0
certifi==2018.1.18
cffi==1.11.5
cfgv==1.0.0
chardet==3.0.4
colorama==0.3.9
colorclass==2.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.1.4
decorator==4.2.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.3
django-allauth==0.35.0
django-cors-headers==2.2.0
django-crispy-forms==1.7.1
django-debug-toolbar==1.9.1
django-extensions==2.0.3
django-filter==1.1.0
django-heroku==0.3.1
django-rest-auth==0.9.3
django-rest-knox==3.1.3
djangorestframework==3.7.7
docopt==0.6.2
factory-boy==2.10.0
Faker==0.8.11
google-api-python-client==1.6.5
gunicorn==19.7.1
httplib2==0.10.3
identify==1.0.7
idna==2.6
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
nodeenv==1.2.0
oauth2client==4.1.2
oauthlib==2.0.6
olefile==0.45.1
packaging==17.1
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pip-upgrader==1.4.4
pre-commit==1.7.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
Pygments==2.2.0
pygsheets==1.1.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
redgreenunittest==0.1.1
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
terminaltables==3.1.0
text-unidecode==1.2
traitlets==4.3.2
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
whitenoise==3.3.1
win-unicode-console==0.5
